+++ keymap.c
@@ -12,7 +12,6 @@ enum {
   MVMNT,
   FUNC,
   NMBR,
-  QWERT,
   PLVR,
   SWPHND,
 };
@@ -72,7 +71,7 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
  *   | PgUp |PgDwn | Alt  |SWPHND| NMBR |                                       | MVMNT|SWPHND|      |      |      |
  *   `----------------------------------'                                       `----------------------------------'
  *                                        ,-------------.       ,-------------.
- *                                        | FUNC | PLVR |       |QWERTY| FUNC |
+ *                                        | FUNC | PLVR |       |      | FUNC |
  *                                 ,------|------|------|       |------+--------+------.
  *                                 |      |      |  F2  |       |  F5  |        |      |
  *                                 | Space|Shift |------|       |------| Enter  | BkSp |
@@ -97,7 +96,7 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
                              KC_H,   KC_N,     KC_E,      KC_I,     KC_O,               KC_QUOT,
              TD(CT_SWITCH),     KC_K,   KC_M,     KC_COMM,   KC_DOT,   KC_SLSH,            TD(CT_UNDER),
                                      MO(MVMNT),TT(SWPHND),KC_NO,    KC_NO,              KC_NO,
-             TT(QWERT),      OSL(FUNC),
+             KC_NO,      OSL(FUNC),
              KC_F5,
              KC_LEAD,         KC_ENT, KC_BSPC
     ),
@@ -236,52 +235,6 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
        KC_TRNS, KC_TRNS,
        KC_TRNS,
        KC_TRNS, KC_TRNS, KC_TRNS
-),
-
-/* Keymap 4: Qwerty
- *
- * ,-----------------------------------------------------.           ,-----------------------------------------------------.
- * |           |      |      |      |      |      |      |           |      |      |      |      |      |      |           |
- * |-----------+------+------+------+------+-------------|           |------+------+------+------+------+------+-----------|
- * |           |  Q   |  W   |   E  |   R  |   T  |      |           |      |  Y   |   U  |   I  |   O  |  P   |           |
- * |-----------+------+------+------+------+------|      |           |      |------+------+------+------+------+-----------|
- * |           |  A   |  S   |   D  |   F  |   G  |------|           |------|  H   |  J   |   K  |   L  |  ;   |           |
- * |-----------+------+------+------+------+------|      |           |      |------+------+------+------+------+-----------|
- * |           |  Z   |  X   |   C  |   V  |   B  |      |           |      |  N   |  M   |   ,  |  .   |   /  |           |
- * `-----------+------+------+------+------+-------------'           `-------------+------+------+------+------+-----------'
- *      |      |      |      |      |      |                                       |      |      |      |      |      |
- *      `----------------------------------'                                       `----------------------------------'
- *                                         ,-------------.           ,-------------.
- *                                         |      |      |           |      |      |
- *                                  ,------|------|------|           |------+------+------.
- *                                  |      |      |      |           |      |      |      |
- *                                  |      |      |------|           |------|      |      |
- *                                  |      |      |      |           |      |      |      |
- *                                  `--------------------'           `--------------------'
- */
-
-[QWERT] = LAYOUT_ergodox(
-// left hand
- KC_TRNS ,KC_TRNS     ,KC_TRNS         ,KC_TRNS          ,KC_TRNS    ,KC_TRNS    ,KC_TRNS
-,KC_TRNS ,KC_Q        ,KC_W            ,KC_E             ,KC_R       ,KC_T      ,KC_TRNS
-,KC_TRNS ,KC_A        ,KC_S            ,KC_D             ,KC_F       ,KC_G
-,KC_TRNS ,KC_Z        ,KC_X            ,KC_C             ,KC_V       ,KC_B     ,KC_TRNS
-,KC_TRNS ,KC_TRNS     ,KC_TRNS         ,KC_TRNS          ,KC_TRNS
-
-                                                                        ,KC_TRNS ,KC_TRNS
-                                                                                 ,KC_TRNS
-                                                          ,KC_TRNS  ,KC_TRNS, KC_TRNS
-
-                                                                // right hand
-                                                               ,KC_TRNS ,KC_TRNS ,KC_TRNS       ,KC_TRNS       ,KC_TRNS          ,KC_TRNS     ,KC_TRNS
-                                                               ,KC_TRNS ,KC_Y    ,KC_U          ,KC_I          ,KC_O             ,KC_P        ,KC_TRNS
-                                                                        ,KC_H    ,KC_J          ,KC_K          ,KC_L             ,KC_SCLN     ,KC_TRNS
-                                                               ,KC_TRNS ,KC_N    ,KC_M          ,KC_COMM       ,KC_DOT           ,KC_SLSH     ,KC_TRNS
-                                                                                 ,KC_TRNS       ,KC_TRNS       ,KC_TRNS          ,KC_TRNS     ,KC_TRNS
-
-                                                               ,KC_TRNS ,KC_TRNS
-                                                               ,KC_TRNS
-                                                               ,KC_TRNS ,KC_TRNS ,KC_TRNS
 ),
 
 
@@ -536,11 +489,6 @@ void matrix_scan_user(void) {
       } else if (layer == NMBR) {
         ergodox_right_led_3_on();
         ergodox_right_led_3_set (LED_BRIGHTNESS_LO);
-      } else if (layer == QWERT) {
-        ergodox_right_led_3_on();
-        ergodox_right_led_3_set (LED_BRIGHTNESS_LO);
-        ergodox_right_led_1_on();
-        ergodox_right_led_1_set (LED_BRIGHTNESS_LO);
       } else if (layer == PLVR) {
         ergodox_right_led_1_on();
         ergodox_right_led_1_set (LED_BRIGHTNESS_LO);